Moves permissions to mainactivity and made locationview only hanlde composable function
